leaderboard row onpress + refactor
